Revert "Move metadata tests from VRT to unit tests"
This reverts commit 48768085d6a993c1f98cf6020dd41a3551c8bd54.
